+++ test/cases/_pug/inline-css-in-style-tag/webpack.config.js
@@ -1,0 +1,32 @@
+const path = require('path');
+const PugPlugin = require('@test/html-bundler-webpack-plugin');
+
+module.exports = {
+  mode: 'production',
+
+  output: {
+    path: path.join(__dirname, 'dist/'),
+  },
+
+  plugins: [
+    new PugPlugin({
+      entry: {
+        index: './src/index.pug',
+      },
+      css: {
+        filename: '[name].[contenthash:8].css',
+      },
+      preprocessor: 'pug',
+      preprocessorOptions: {},
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader', 'sass-loader'],
+      },
+    ],
+  },
+};